Extract infrastructure layer (#251).
