+++ cruftbot/infrastructure/settings/components/activity_stream.py
@@ -1,0 +1,4 @@
+from cruftbot.infrastructure.settings.components.base import INSTALLED_APPS
+
+
+INSTALLED_APPS.append("actstream")
