+++ cruftbot/infrastructure/settings/components/debug_toolbar.py
@@ -1,7 +1,7 @@
 from debug_toolbar.settings import PANELS_DEFAULTS
 
-from cruftbot.settings.components.base import INSTALLED_APPS
-from cruftbot.settings.components.base import MIDDLEWARE
+from cruftbot.infrastructure.settings.components.base import INSTALLED_APPS
+from cruftbot.infrastructure.settings.components.base import MIDDLEWARE
 
 
 INSTALLED_APPS.append("debug_toolbar")
@@ -13,7 +13,7 @@ MIDDLEWARE.append("debug_toolbar.middleware.DebugToolbarMiddleware")
 DEBUG_TOOLBAR_PANELS = PANELS_DEFAULTS + ["stories_django.debug_toolbar.StoriesPanel"]
 
 DEBUG_TOOLBAR_CONFIG = {
-    "SHOW_TOOLBAR_CALLBACK": "cruftbot.settings.development.show_toolbar"
+    "SHOW_TOOLBAR_CALLBACK": "cruftbot.infrastructure.settings.development.show_toolbar"
 }
