+++ cruftbot/infrastructure/settings/components/extra_checks.py
@@ -1,4 +1,4 @@
-from cruftbot.settings.components.base import INSTALLED_APPS
+from cruftbot.infrastructure.settings.components.base import INSTALLED_APPS
 
 
 INSTALLED_APPS.append("extra_checks")
