+++ cruftbot/infrastructure/settings/components/migration_linter.py
@@ -1,0 +1,4 @@
+from cruftbot.infrastructure.settings.components.base import INSTALLED_APPS
+
+
+INSTALLED_APPS.append("django_migration_linter")
