+++ cruftbot/infrastructure/settings/components/querycount.py
@@ -1,0 +1,4 @@
+from cruftbot.infrastructure.settings.components.base import MIDDLEWARE
+
+
+MIDDLEWARE.append("querycount.middleware.QueryCountMiddleware")
