+++ cruftbot/infrastructure/settings/components/test_migrations.py
@@ -1,4 +1,4 @@
-from cruftbot.settings.components.base import INSTALLED_APPS
+from cruftbot.infrastructure.settings.components.base import INSTALLED_APPS
 
 
 INSTALLED_APPS.append("django_test_migrations.contrib.django_checks.AutoNames")
