+++ cruftbot/infrastructure/urls/__init__.py
@@ -3,8 +3,8 @@ import allauth.urls
 from django.urls import include
 from django.urls import path
 
-from cruftbot.views.index import index
-from cruftbot.views.schema import schema_view
+from cruftbot.infrastructure.views.index import index
+from cruftbot.infrastructure.views.schema import schema_view
 
 urlpatterns = [
     path("accounts/", include(allauth.urls)),
