+++ cruftbot/infrastructure/urls/development.py
@@ -2,7 +2,7 @@ import debug_toolbar
 from django.urls import include
 from django.urls import path
 
-from cruftbot.urls import urlpatterns
+from cruftbot.infrastructure.urls import urlpatterns
 
 
 development_patterns = [path("__debug__/", include(debug_toolbar.urls))]
