+++ cruftbot/infrastructure/urls/production.py
@@ -1,0 +1,4 @@
+from cruftbot.infrastructure.urls import urlpatterns
+
+
+__all__ = ["urlpatterns"]
