+++ cruftbot/infrastructure/views/schema.py
