--- cruftbot/settings/components/activity_stream.py
@@ -1,4 +1,0 @@
-from cruftbot.settings.components.base import INSTALLED_APPS
-
-
-INSTALLED_APPS.append("actstream")
