--- cruftbot/settings/components/querycount.py
@@ -1,4 +1,0 @@
-from cruftbot.settings.components.base import MIDDLEWARE
-
-
-MIDDLEWARE.append("querycount.middleware.QueryCountMiddleware")
