--- cruftbot/urls/production.py
@@ -1,4 +1,0 @@
-from cruftbot.urls import urlpatterns
-
-
-__all__ = ["urlpatterns"]
